refactor: Remove separate default `DetectorElementBase` file
`DetectorElementBase.hpp` has an `ifdef` that either includes the
preprocessor define or previously included the default base class
header. We included that header in a numbr of locations by accident,
breaking the pattern. This removes the separate header (and erroneous
inlcudes) and puts the default interface into `DetectorElementBase.hpp`
inside the `ifdef`.
