+++ Core/src/Geometry/ProtoLayer.cpp
@@ -10,7 +10,6 @@
 
 #include "Acts/Definitions/Algebra.hpp"
 #include "Acts/Geometry/Polyhedron.hpp"
-#include "Acts/Geometry/detail/DefaultDetectorElementBase.hpp"
 #include "Acts/Utilities/Helpers.hpp"
 
 #include <algorithm>
