+++ Core/src/Surfaces/Surface.cpp
@@ -10,7 +10,6 @@
 
 #include "Acts/Definitions/Common.hpp"
 #include "Acts/EventData/detail/TransformationBoundToFree.hpp"
-#include "Acts/Geometry/detail/DefaultDetectorElementBase.hpp"
 #include "Acts/Surfaces/SurfaceBounds.hpp"
 #include "Acts/Surfaces/detail/AlignmentHelper.hpp"
 #include "Acts/Utilities/VectorHelpers.hpp"
