+++ Plugins/Geant4/include/Acts/Plugins/Geant4/Geant4DetectorElement.hpp
@@ -11,7 +11,6 @@
 #include "Acts/Definitions/Algebra.hpp"
 #include "Acts/Geometry/DetectorElementBase.hpp"
 #include "Acts/Geometry/GeometryContext.hpp"
-#include "Acts/Geometry/detail/DefaultDetectorElementBase.hpp"
 #include "Acts/Surfaces/Surface.hpp"
 
 #include <memory>
