+++ Tests/CommonHelpers/Acts/Tests/CommonHelpers/CubicTrackingGeometry.hpp
@@ -21,7 +21,6 @@
 #include "Acts/Geometry/PlaneLayer.hpp"
 #include "Acts/Geometry/TrackingGeometry.hpp"
 #include "Acts/Geometry/TrackingVolume.hpp"
-#include "Acts/Geometry/detail/DefaultDetectorElementBase.hpp"
 #include "Acts/Material/HomogeneousSurfaceMaterial.hpp"
 #include "Acts/Surfaces/PlaneSurface.hpp"
 #include "Acts/Surfaces/RectangleBounds.hpp"
